[libclang/python] Add type annotations
The usage of the libclang's Python binding allows users to learn and
explore the clang library. A critical part of being able to easily
explore this interface is having proper Python type annotations, that an
IDE and an interactive repl can understand.

Without this proper support, it is difficult to interactively traverse
a Cursor tree or write code with IDE support.

https://discourse.llvm.org/t/type-annotations-for-libclang-python-bindings/70644